fix: use int for ord_cam to avoid float suffix breaking singola commessa search
Also documents near-zero press times and negative ORE_COTT cases in data integrity notes.

+++ app/_pages/reports.py
@@ -66,7 +66,7 @@ def render(report_name: str) -> None:
             run = False
         else:
             try:
-                params["ord_cam"] = float(ord_str)
+                params["ord_cam"] = int(ord_str)
             except ValueError:
                 st.error("Ord. Camfart deve essere un numero.")
                 return
